moved core game logic out of game_reducer and into external function file
